pyproject.toml
[project]
name = "ai-assistant-service"
version = "0.1.0"
description = "HTTP-based AI assistant service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.34.3",
    "pydantic>=2.4.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "lightspeed-stack>=0.3.0",
    "prometheus-client>=0.22.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true